Move map components and pages
