+++ apps/client/src/features/map/components/UserDot.jsx
