+++ apps/client/src/features/map/pages/Home.jsx
@@ -21,15 +21,15 @@ import LoginIcon from "@mui/icons-material/Login";
 import DarkModeIcon from "@mui/icons-material/DarkMode";
 import LightModeIcon from "@mui/icons-material/LightMode";
 import { useNavigate } from "react-router-dom";
-import GlobalLoadingProgress from "../components/GlobalLoadingProgress";
+import GlobalLoadingProgress from "../../../components/GlobalLoadingProgress";
 import LocateUser from "../components/LocateUser";
 import VendorMarkers from "../components/VendorMarkers";
 import CreditMarker from "../components/CreditMarker";
 import RoutingEngine from "../components/RoutingEngine";
-import useAuth from "../hooks/useAuth";
-import useLogout from "../hooks/useLogout";
+import useAuth from "../../../hooks/useAuth";
+import useLogout from "../../../hooks/useLogout";
 import { useThemeStore } from "@/store/themeStore";
-import "../assets/map.css";
+import "../../../assets/map.css";
 
 const INITIAL_VIEW_STATE = {
   longitude: 21.409471852749466,
